Full Deployment Shuffle Stage 1
= Moved Backend Express server to root level of project repository
= Adjusted the frontend package.json to support netlify web-builds
