+++ backend-express-app/models/mongodb_schemas/Household.js
