+++ backend-express-app/routes/UpdateUserEmail.js
@@ -16,5 +16,3 @@ app.put('/users/:id/email', (req, res) => {
 
       }
   });
-
-  module.exports.handler = serverless(router);
