+++ backend-express-app/routes/listings.js
@@ -3,7 +3,6 @@ const jwt = require("jsonwebtoken");
 const router = express.Router();
 const Listing = require("../models/listing.model");
 const User = require("../models/user.model");
-const serverless = require("serverless-http");
 
 router.post("/add", async (req, res) => {
   const token = req.cookies.token || req.body.token;
@@ -233,4 +232,4 @@ router.post("/delete", async (req, res) => {
   }
 });
 
-module.exports.handler = serverless(router);
+module.exports = router;
