+++ backend-express-app/routes/updateUser.js
@@ -3,9 +3,8 @@ const jwt = require('jsonwebtoken');
 const router = express.Router();
 const User = require('../models/user.model');
 const Listing = require('../models/listing.model');
-const serverless = require("serverless-http");
-const app = express();
 
+//Might need to use post instead of patch
 router.patch('/user', async (req, res) => {
     const token = req.cookies.token || req.body.token;
     // Check if user is logged in
@@ -21,6 +20,7 @@ router.patch('/user', async (req, res) => {
       const update = {
         first_name,
         last_name,
+		zip_code,
         gender,
         email,
 		    desc
@@ -43,16 +43,24 @@ router.patch('/user', async (req, res) => {
         }
         return res.status(200).json({ message: 'OK' });
       } catch (err) {
-        console.log('Error:', err);
-        return res.status(400).json({ message: err.message })
+		//Validation failures happen here
+		if(!err.path){
+			
+			const errorField = Object.keys(err.errors)[0];
+			
+			console.log('Error Database Failed Validation:', errorField);
+			return res.status(400).json({ message: errorField })
+		}
+		else{
+			console.log('Error Database Failed To Post:', err.path);
+			return res.status(400).json({ message: err.path })
+		}
       }    
     } catch (error) {
       // If the token is invalid or has expired, return a 401 Unauthorized response
+	  console.log('Schema Error', error);
       return res.status(403).json({ message: 'Request error on submission' });
     }
 });
 
-app.use(express.json());
-app.use('/.netlify/functions/updateUser', router);
-
-module.exports.handler = serverless(app);
+module.exports = router;
