+++ backend-express-app/routes/users.js
@@ -1,10 +1,5 @@
 const router = require('express').Router();
 let User = require('../models/user.model');
-const serverless = require("serverless-http");
-
-router.get("/", (req, res) => {
-    res.json({'path':'App is running!'});
-  });
 
 router.route('/').get((req, res) => {
     User.find()
@@ -21,4 +16,6 @@ router.route('/add').post((req, res) => {
         .catch(err => res.status(400).json('Error: '+ err));
 });
 
-module.exports.handler = serverless(router);
+
+
+module.exports = router;
