--- react-native-app/backend-express-app/netlify.toml
@@ -1,4 +1,0 @@
-[build]
-    functions = "functions"
-    command = "npm run build"
-    publish = "dist"
