+++ react-native-app/package.json
@@ -6,7 +6,9 @@
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web"
+    "web": "expo start --web",
+    "webbuild": "expo export:web"
+	
   },
   "dependencies": {
     "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
